Solves the thread related problem of calling FreeLibrary ()
The previous exit problem was due to the lack of execution of MailSlot,
that is to say, threads were exited and their exit mechanisms were
intact. A more fundamental reason is that no virtual destructor is used.
The solution is simple: add a virtual, plus a virtual. increase in
defense